requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.7
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
spacy==3.7.2
rapidfuzz==3.5.2

